+++ src/views/ItemView.js
@@ -714,6 +714,15 @@ class ItemView extends React.Component {
                   />
                 </span>
               )}
+              paneSub={(
+                <FormattedMessage
+                  id="ui-inventory.instanceRecordSubtitle"
+                  values={{
+                    hrid: item?.hrid,
+                    updatedDate: getDate(item?.metadata?.updatedDate),
+                  }}
+                />
+              )}
               dismissible
               onClose={this.props.onCloseViewItem}
               actionMenu={this.getActionMenu}
@@ -836,10 +845,14 @@ class ItemView extends React.Component {
                   />
                   {' '}
                   <FormattedMessage
-                    id="ui-inventory.itemRecordWithDescription"
+                    id={
+                      item?.isBoundWith ?
+                        'ui-inventory.itemRecordWithDescriptionBW' :
+                        'ui-inventory.itemRecordWithDescription'
+                    }
                     values={{
                       materialType: item?.materialType?.name,
-                      status: item?.status?.name
+                      status: item?.status?.name,
                     }}
                   />
                 </Col>